requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
environs==8.0.0
marshmallow==3.6.1
psycopg2-binary==2.8.5
pycparser==2.20
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
